Deal with division by zero in precision_recall()

+++ sp_model.py
@@ -46,23 +46,39 @@ def build_model(input_shape=(64, 64, 3), dropout_ratio=0.3, convolution_layers=(
     return model
 
 
-def precision_recall(y_pred, y_label):
+def precision_recall(y_pred, y_label, verbose=True):
     tp = sum([1 for pred, true in zip(y_pred, y_label) if pred == true == 1])
     fp = sum([1 for pred, true in zip(y_pred, y_label) if pred != true and true == 0])
     tn = sum([1 for pred, true in zip(y_pred, y_label) if pred == true == 0])
     fn = sum([1 for pred, true in zip(y_pred, y_label) if pred != true and true == 1])
 
-    precision = float(tp) / (tp + fp)
-    recall = float(tp) / (tp + fn)
-    f1_score = 2. / (1. / recall + 1. / precision)
+    if (tp + fp) == 0:
+        precision = float('inf')
+    else:
+        precision = float(tp) / (tp + fp)
+    if (tp + fn) == 0:
+        recall = float('inf')
+    else:
+        recall = float(tp) / (tp + fn)
 
-    print('''True positive: {}
+    if recall == float('inf') and precision == float('inf'):
+        f1_score = float('inf')
+    elif recall == float('inf'):
+        f1_score = 2. * precision
+    elif precision == float('inf'):
+        f1_score = 2. * recall
+    else:
+        f1_score = 2. / (1. / recall + 1. / precision)
+
+    if verbose:
+        print('''True positive: {}
 True negative: {}
 False positive: {}
 False negative: {}
 Precision: {}
 Recall: {}
 F1-Score: {}'''.format(tp, tn, fp, fn, precision, recall, f1_score))
+    return precision, recall, f1_score
 
 
 def model_predict(model, filenames_csv, predict_folder, fullsize_folder, output_folder, image_size, label_map,
@@ -75,7 +91,7 @@ def model_predict(model, filenames_csv, predict_folder, fullsize_folder, output_
                                            featurewise_std_normalization=True)
 
     y_predict = model.predict(x_all)
-    y_predict_with_bias = y_predict
+    y_predict_with_bias = np.copy(y_predict)
     if positive_bias:
         y_predict_with_bias[:, 0] -= positive_bias
         y_predict_with_bias[:, 1] += positive_bias
@@ -114,7 +130,7 @@ def model_predict(model, filenames_csv, predict_folder, fullsize_folder, output_
 
 def model_saliency(model, filenames_csv, predict_folder, fullsize_folder, output_folder, image_size, full_image_size,
                    layer_idx=-1, filter_indices=0, exclude_index=4, save_sal=False, save_overlay=True,
-                   print_prediction=True):
+                   print_prediction=True, skip_existing=True):
     test_img, test_labels = sp_googlemaps.load_data(filenames_csv, predict_folder, image_size, bool, exclude_index,
                                                     featurewise_center=True,
                                                     featurewise_std_normalization=True)
@@ -135,6 +151,11 @@ def model_saliency(model, filenames_csv, predict_folder, fullsize_folder, output
         filename = f[0]
         result_filename = os.path.join(output_folder, 'sal_' + filename)
         overlay_result_filename = os.path.join(output_folder, 'sal_ovl_' + filename)
+        if skip_existing and (
+                (not save_overlay or os.path.exists(overlay_result_filename))
+                and (not save_sal or os.path.exists(result_filename))
+        ):
+            continue
         sal = visualize_saliency(model, layer_idx=layer_idx, filter_indices=filter_indices,
                                  seed_input=test_img[i, :, :, :])
         imageio.imwrite(result_filename, sal)
